changes to escript/py_src/pdetools.py and /escript/src/Data.h/.cpp to make the Locator work in MPI. escript::Data::mindp now returns a 3 tuple, with the MPI rank of the process on which the minimum value occurs included. escript::Data::convertToNumArrayFromDPNo also takes the ProcNo to perform the MPI reduction.
This had to be implemented in both the MPI and non-MPI versions to allow
the necesary changes to the Python code in pdetools.py. In the non-MPI
version ProcNo is set to 0. This works for the explicit scripts tested
thus far, however if it causes problems in your scripts contact Ben or
Lutz, or revert the three files (pdetools.py, Data.h and Data.cpp) to
the previous version.  



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@790 77569008-7704-0410-b7a0-a92fef0b09fd