Merge of branches/windows_from_1431_trunk.
Revamp of the exception system.
Fix unused vars and signed/unsigned comparisons.
defined a macro THROW(ARG) in the system_dep.h's to
deal with the expectations of declarations on different architectures.

Details in the logs of branches/windows_from_1431_trunk.

pre-merge snapshot of the trunk in tags/trunk_at_1452



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@1455 77569008-7704-0410-b7a0-a92fef0b09fd