This is a fake commit, to note that
Finley_IndexList_insertIndex is being called from paso's Pattern.c.

I was under the impresssion that this was a violation of the module rule forbidding
paso from calling Finley code. I have not fixed this, since I am not the
best person to formulate the workaround. My immediate suggestion would be to move the
functionality into paso, and call it from Finley. However, I'm sure there is a better way.



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@1916 77569008-7704-0410-b7a0-a92fef0b09fd