I have changed the default scons options for gcc to be much more picky about warnings. By default warnings will be treated as errors; so a program with warnings won't compile.
If you need to turn this off, usewarnings=no will work.

The previous usepedantic oprtion is no longer turned on by default.

This should have no effect on any of the other compilers.
If anyone wants to jump in and supply sensible values for the fatalwarnings variable for their favourite compiler go ahead.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@2026 77569008-7704-0410-b7a0-a92fef0b09fd