I may get into trouble for this. boost-python 1.34 does have a docstring_options class, but does not have a 3 argument constructor for it. So the test has been modified to
#if ((BOOST_VERSION/100)%1000 > 34) || (BOOST_VERSION/100000 >1)

If you wish to make things more delicate, one can define a 2 argument construction
of docopt just for 1.34 (with an #elif). Probably not worth the effort frankly.

Hope that this has not broken anything for anyone else. The SVN logs suggest this is a
little fragile.....

Also, please be aware that much of our chemistry interface code, that we wish to use with
escript, makes extensive use of boost python.
Having two different boost versions mucking with the python interpreter sounds
like a really bad idea, I'm sure you'll agree.
The problem is that it is not a simple task for us to build new versions of boost-python
on all our platforms. Consequently, it would be nice to be informed when you guys
intend to upgrade a support library of this nature so that we can plan and allocate
resources to keep up.

Cheers.



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@2076 77569008-7704-0410-b7a0-a92fef0b09fd