inf, sup and Lsup now correctly handle +-infinity. They also will return NaN if any part of their input is NaN.
This will break unit tests since it exposes the hidden bug (#447 in mantis)

This code relies on the ability to test for NaNs.
To do this it makes use of macros and functions from C99.
If you do not have a C99 compiler, then you will probably get the old behaviour.
That is, you won't know when you have NaNs.


Also did minor tweak to saveDataCSV doco.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@2766 77569008-7704-0410-b7a0-a92fef0b09fd