Updated various nan checks to consider the windows _isnan
The default compiler flags have changed as well.
+ intel will now take -std=c99 instead of -ansi
+ gcc has -ansi removed which means it defaults to gnu99

We could have set gcc to -std=c99 as well but that gives a 
warning on g++ which gets converted into an error by our
pedantic warning.

Rationale:
We need something more than ansi to get proper nan handling.
- We currently don't have any code which does not comply with ansi
  but the nan checks don't work.

Impact:
If we want our code to still be able to compile on older compilers
(at reduced functionality) we need to be careful not to introduce other
c99-isms.
If we don't care, then it's time for some celebratory // comments.



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@2884 77569008-7704-0410-b7a0-a92fef0b09fd