Added/modified AMG tests:
  - "self.failUnless(...)" became "self.assertTrue(...)"
    because fail* is soon to be deprecated,
  - fixed logical error in self.assertTrue argument,
  - added another test, 'testSquare'.

Added some print statements to AMG.c in verbose mode to
identify which of the stopping conditions were fulfilled.

Changed default AMG parameters for 'setNumPreSweeps' and 
'setNumPostSweeps' to 1 (huge speed improvment).

P.S Beware of bugs in the above code; I have only proved 
it correct, not tried it... Jokes!



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@3436 77569008-7704-0410-b7a0-a92fef0b09fd