Hopefully, this will address the interpolation problems.
New canInterpolate() function exposed to python which calls probeInterpolation

AbstractDomain now has an additional virtual.
preferredIntrpolationOnDomain()

This will return 0 if interpolation is impossible, 1 if possible and preferred.
It will return -1 if interpolation is possible and preferred in the 
oposite direction.

A value of -1 does not say that the proposed interpolation is possible or not.
Rather it indicates "please use the other way".
If you really _need_ to test it that way, use probeInterpolationOnDomain 



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@4255 77569008-7704-0410-b7a0-a92fef0b09fd