Some reorganisation of EscriptParams and related changes. We now distinguish between modifiable parameters and build features. To interrogate the latter use: escript::hasFeature()  [ python: escript.hasFeature() ] and escript::listFeatures() [ python: escript.listFeatures() ]
I have decided to add specific getters for the remaining few parameters to
avoid the penalty of string comparisons and to remove the wealth of friend
declarations with Data* classes.

Also made some changes in SConstruct to make sure we set *all* preprocessor
directives *before* building anything.

Still to do is dealing with checks for direct solver and gmsh. The latter needs
to be a runtime check rather than a compile time check.


git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@6174 77569008-7704-0410-b7a0-a92fef0b09fd