change to a cppunit assertion which currently fails on guineapig to make it output a sane error.
The reason for the failure is not entirely clear.
We are closing a parallel file which, according to doco, is a collective
operation, however the subsequent check for file size almost certainly returns
the wrong value on at least one rank.
This could be easily fixed by a barrier but we never saw this behaviour so
I suggest we investigate.


git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@6227 77569008-7704-0410-b7a0-a92fef0b09fd