I have identified and fixed the bug that was causing the dudley test run_inputOutput.py and the finley test run_inputOutput.py to fail.
This error occured because the environment variables DUDLEY_TEST_DATA and FINLEY_TEST_DATA were not defined. As a result, the test script was trying to load files from a directory that didn't exist, causing the functions readGmsh and readMesh to bounce back an error. I have modified these tests so that when DUDLEY_TEST_DATA and FINLEY_TEST_DATA are undefined, the script defaults to the correct directory. This has fixed the error and the test now runs to completion.



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@6733 77569008-7704-0410-b7a0-a92fef0b09fd