Minor bug fixes.
I fixed several errors that were preventing eScripts from compiling (on Debian Stretch) when CUDA was enabled.

A set of problems arose due to a preprocessor directive starting on line 869 of RipleyDomain.cpp that the compiler only includes in the specific case where CUDA is enabled but MPI is not. Specifically:
(i) the switch statement inside this directive did not explicitly handle several values of the enum variables in escript/SolverOptions.h and had no default case. As a result the compiler was bouncing back an error. I have resolved this issue by adding a default case to the statement.
(ii) the if statement following the switch in this preprocessor directive set the variable "package" to a non-existent enum "SO_PACKAGE_DEFAULT" which I believe is a typo for "SO_DEFAULT".

I have also fixed:
(iii) The file ripley_UnitTest.cpp was trying to include the nonexistent file "escript/Esys_MPI.h" instead of "escript/EsysMPI.h", causing the compiler to bounce back an error.




git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@6736 77569008-7704-0410-b7a0-a92fef0b09fd