Fixed bug 450 (saveSilo tests were failing).
The tests were failing on meerkat because Python could not locate the example mesh files when WEIPA_WORKDIR was undefined. This is similar to the problem that was causing the dudley test run_inputOutput.py to fail.

While debugging, I noticed that the saveSilo test script was creating a lot of files and then not deleting them after the tests had run. I have changed a couple of lines to fix this.



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@6748 77569008-7704-0410-b7a0-a92fef0b09fd