I have fixed the bug that was causing the tests run_gravity.py and run_inversion_gravmag_2d.py to fail on the buildbot gnu-omp-py3-paso-long.
The interpolation function I coded for minimizers.py finds the minimum of a parabola that goes through the points phi(alpha_lo) and phi(alpha_hi) and whose gradient matches phi(alpha) at alpha_lo. Occasionally the minimum of this function lies outside of the domain [alpha_lo, alpha_hi] and, when this occurs, the line search function breaks down. The interpolation function now checks that the minimum is within the interval and if it is not it falls back on the bisection method.



git-svn-id: https://svn.geocomp.uq.edu.au/svn/esys13/trunk@6766 77569008-7704-0410-b7a0-a92fef0b09fd